32->64?

+++ CHAPPIE/tests/test_infrastructure_assets.py
@@ -55,10 +55,10 @@ def test_get_levees():
 
     expected_file = os.path.join(EXPECTED_DIR, 'levees.parquet')
     expected = geopandas.read_parquet(expected_file)
-    
+
     field_list = ['SYSTEM_ID', 'LEVEED_ID', 'SEG_COUNT',
                   'Shape__Area', 'Shape__Length']
     for col in field_list:
-        expected[col] = expected[col].astype('int32')
+        expected[col] = expected[col].astype('int64')
 
     assert_geodataframe_equal(actual, expected)
